migration to fastendpoints

+++ src/Application/Queries/Orders/Statistics/GetOrderStatisticQueryResponse.cs
@@ -1,11 +1,11 @@
 ﻿namespace Application.Queries.Orders.Statistics;
 public readonly record struct GetOrderStatisticQueryResponse(IEnumerable<MerchItemStatistic> Statistics,
-                                                    int AbsoluteAmountSold,
-                                                    int AbsolutePrice,
-                                                    int AbsoluteSelfPrice);
+                                                             int AbsoluteAmountSold,
+                                                             decimal AbsolutePrice,
+                                                             decimal AbsoluteSelfPrice);
 
 public readonly record struct MerchItemStatistic(string MerchItemName,
-                                        int OrdersWithItemCount,
-                                        int TotalSelfPrice,
-                                        int TotalPrice,
-                                        int TotalAmount);
+                                                 int OrdersWithItemCount,
+                                                 decimal TotalSelfPrice,
+                                                 decimal TotalPrice,
+                                                 decimal TotalAmount);
